feat: dto 추가 및 수정

+++ src/main/java/com/nhnacademy/taskapi/order/entity/OrderStatus.java
@@ -13,7 +13,7 @@ public class OrderStatus {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "order_status_id")
-    private Long orderStatusId;
+    private int orderStatusId;
 
     @UniqueElements
     @Column(name = "order_status_name")
